requirements.txt
cm-client
pyyaml
